rename upload docs
