+++ src/applications/disability-benefits/all-claims/pages/uploadPersonalPtsdDocuments.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import fileUploadUI from 'us-forms-system/lib/js/definitions/file';
 import environment from '../../../../platform/utilities/environment';
 
-import { DocumentDescription } from '../content/uploadPtsdDocuments';
+import { DocumentDescription } from '../content/uploadFormDocuments';
 import { PtsdNameTitle } from '../content/ptsdClassification';
 
 const FIFTY_MB = 52428800;
